study26/app2/pyproject.toml
[project]
name = "app2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.2",
    "httpx>=0.28.1",
    "kafka-python>=2.3.0",
    "pydantic-settings>=2.12.0",
]
